fix length type
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@4021 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/HelpViewer/Parser.m
@@ -36,13 +36,12 @@
 + (void) parserWithHandler: (id<SAXHandler>) handler
     withData: (NSData*) data
 {
-    //NSString* file = [NSString stringWithContentsOfFile: filename];
     NSString* file = [[NSString alloc] initWithData: data encoding: NSISOLatin1StringEncoding];
     NSMutableString* current = [[NSMutableString alloc] init];
 
     if (file != nil)
     {
-	int i;
+	NSUInteger i;
 	unsigned entity = 0;
 	char entityBuf[MAX_ENTITY_LEN+1];
 	BOOL Tag = NO;
